# swagger/swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Documentation",
    "description": "This is the API documentation for VMess project",
    "version": "1.0.0"
  },
  "host": "localhost:1002",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng ký tài khoản mới",
        "description": "Endpoint để đăng ký tài khoản mới với email và mật khẩu.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/verifyCode": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Xác thực mã code",
        "description": "Endpoint để Xác thực mã code.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng nhập",
        "description": "Endpoint để đăng nhập với email và mật khẩu.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/checkToken": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Kiểm tra Access Token",
        "description": "Endpoint để kiểm tra Access Token có hợp lệ hay không khi mới vào trang.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/refreshToken": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Làm mới Access Token",
        "description": "Endpoint để lấy Access Token mới bằng Refresh Token.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng xuất",
        "description": "Endpoint để đăng xuất và xóa Refresh Token.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/googleLogin": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng nhập bằng Google",
        "description": "Endpoint để đăng nhập thông qua tài khoản Google.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "tokenId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/forgotPassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Yêu cầu quên mật khẩu",
        "description": "Endpoint để yêu cầu quên mật khẩu",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/verifyResetCode": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Xác thực mã quên mật khẩu",
        "description": "Endpoint để Xác thực mã trong chức năng quên mật khẩu",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/resetPassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đổi mật khẩu mới trong chức năng quên mật khẩu",
        "description": "Endpoint Đổi mật khẩu mới trong chức năng quên mật khẩu",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "newPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}